Extract CactusWeakRef to a weak module
Begin rename to the same struct names as std::rc.
